15.01.2021

+++ components/bootloader_support/src/bootloader_common.c
@@ -23,6 +23,8 @@
 #include "esp32s2/rom/spi_flash.h"
 #elif CONFIG_IDF_TARGET_ESP32S3
 #include "esp32s3/rom/spi_flash.h"
+#elif CONFIG_IDF_TARGET_ESP32C3
+#include "esp32c3/rom/spi_flash.h"
 #endif
 #include "esp_rom_crc.h"
 #include "esp_rom_gpio.h"
@@ -86,7 +88,7 @@ bool bootloader_common_label_search(const char *list, char *label)
 
         // [start_delim] + label + [end_delim] was not found.
         // Position is moving to next delimiter if it is not the end of list.
-        int pos_delim = strcspn(sub_list_start_like_label, ", ");
+        size_t pos_delim = strcspn(sub_list_start_like_label, ", ");
         if (pos_delim == strlen(sub_list_start_like_label)) {
             break;
         }
